+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Add the selection of effect value initialization upon entering effect id
 - [#136] - Add instrument bank (thanks [@papiezak], [@jpcima])
+- Keep selected file filter in instrument and bank import
 
 ### Fixed
 - [#143] - Fix the corruption of note delay effect without note on (thanks [@ImATrackMan])